Updates
Use std::function in interface for Nelder-Mead. This makes it much
easier to define functions on the fly to optimize; e.g., we can use
lambda functions and capture other data in score as part of the
function definition. This is what we are used to doing in R, since
functions there are regular objects.

Got the CustomLinearRegion example running. It appears to work just
about correctly!